line/requirements.txt
fastapi
uvicorn
line-bot-sdk
python-dotenv
python-multipart
pyngrok
google-generativeai